+++ src/main/java/subway/line/presentation/dto/LineResponse.java
@@ -1,4 +1,4 @@
-package subway.Line.presentation.dto;
+package subway.line.presentation.dto;
 
 public class LineResponse {
