+++ src/main/java/subway/line/presentation/dto/SectionRequest.java
@@ -1,19 +1,16 @@
-package subway.Line.presentation.dto;
+package subway.line.presentation.dto;
 
-public class LineRequest {
-
-    private String name;
-    private String color;
+public class SectionRequest {
     private Long upStationId;
     private Long downStationId;
     private int distance;
 
-    public String getName() {
-        return name;
-    }
+    public SectionRequest() {}
 
-    public String getColor() {
-        return color;
+    public SectionRequest(Long upStationId, Long downStationId, int distance) {
+        this.upStationId = upStationId;
+        this.downStationId = downStationId;
+        this.distance = distance;
     }
 
     public Long getUpStationId() {
