+++ src/main/java/subway/path/presentation/PathController.java
@@ -4,11 +4,9 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import subway.Line.presentation.dto.SectionResponse;
 import subway.Station.domain.Station;
 import subway.Station.presentation.dto.StationResponse;
 import subway.path.application.PathService;
