+++ src/test/java/nextstep/subway/unit/LineServiceTest.java
@@ -3,18 +3,15 @@ package nextstep.subway.unit;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-import subway.Line.application.LineService;
-import subway.Line.domain.Line;
-import subway.Line.domain.Sections;
-import subway.Line.infrastructure.LineRepository;
-import subway.Line.presentation.dto.SectionRequest;
-import subway.Station.domain.Station;
+import subway.line.application.LineService;
+import subway.line.domain.Sections;
+import subway.line.infrastructure.LineRepository;
+import subway.line.presentation.dto.SectionRequest;
 import subway.Station.infrastructure.StationRepository;
 import subway.SubwayApplication;
 import subway.global.exception.CustomException;
