+++ src/test/java/nextstep/subway/unit/LineTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import subway.Line.domain.Line;
+import subway.line.domain.Line;
 import subway.Station.domain.Station;
 
 public class LineTest {
