+++ src/test/java/nextstep/subway/unit/PathServiceMockTest.java
@@ -4,17 +4,15 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import java.util.Optional;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
-import subway.Line.domain.Line;
-import subway.Line.domain.Section;
-import subway.Line.domain.Sections;
-import subway.Line.infrastructure.LineRepository;
-import subway.Line.infrastructure.SectionRepository;
+import subway.line.domain.Line;
+import subway.line.domain.Section;
+import subway.line.infrastructure.LineRepository;
+import subway.line.infrastructure.SectionRepository;
 import subway.Station.domain.Station;
 import subway.Station.infrastructure.StationRepository;
 import subway.path.application.PathService;
