+++ src/test/java/subway/line/LineAcceptanceFixture.java
@@ -1,4 +1,4 @@
-package subway.Line;
+package subway.line;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
